libs/backend/guards/src/lib/auth/jwt-auth.guard.ts
import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
import { AuthGuard } from '@nestjs/passport';

@Injectable()
export class JwtAuthGuard extends AuthGuard('jwt') {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  override handleRequest(_: any, user: any) {
    // Here we have the user from the strategy (the real user)
    // throw an exception if user not exists
    if (!user) {
      throw new HttpException('UNAUTHORIZED', HttpStatus.UNAUTHORIZED);
    }

    return user;
  }
}
